requirements.txt
pyyaml
psutil
deprecated
tqdm
azure.cli.core
azure-storage-blob==2.1.0
